pyproject.toml
[tool.poetry]
name = "ws-homework"
version = "0.1.0"
description = ""
authors = ["thediogen <dvodnenkoevgen1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
itsdangerous = "^2.2.0"
uvicorn = "^0.34.0"
websockets = "^14.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
